requirements.txt
requests
pillow
tweepy
